microservice/requirements.txt
Flask==2.1.0
google-cloud-datastore==2.4.0